pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "numic"
version = "0.1.0"
description = "PHVD (post-haemorrhagic ventricular dilation) risk scoring"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "numpy>=1.26",
    "pydicom>=2.4",
    "Pillow>=10.0",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[project.scripts]
numic-manual-measure = "numic.cli.measure:main"

[tool.hatch.build.targets.wheel]
packages = ["src/numic"]

[tool.hatch.build.targets.sdist]
include = ["src/numic"]

[tool.uvicorn]
factory = false
